返回翻页形势

+++ CronJobs/Infrastructure/Model/Response/BasePageResponse.cs
@@ -8,7 +8,7 @@ namespace Infrastructure.Model.Response
         /// <summary>
         /// 总页数
         /// </summary>
-        public Int32 TotalCount { get; set; }
+        public long TotalCount { get; set; }
         /// <summary>
         /// 当前页数据集合
         /// </summary>
